+++ src/app/map/geostories/[geostory_id]/page.tsx
@@ -1,5 +1,7 @@
 'use client';
 
+import { useEffect } from 'react';
+
 import Link from 'next/link';
 
 import type { NextPage } from 'next';
@@ -9,6 +11,7 @@ import type { MonitorParsed } from '@/types/monitors';
 
 import { useGeostoryLayers } from '@/hooks/geostories';
 import { useMonitors } from '@/hooks/monitors';
+import { useSyncLayersSettings, useSyncCompareLayersSettings } from '@/hooks/sync-query';
 
 import DatasetCard from '@/components/datasets/card';
 import GeostoryHead from '@/components/geostories/header';
@@ -25,8 +28,21 @@ const GeostoryPage: NextPage<{ params: { geostory_id: string } }> = ({
   const { data, isLoading, isFetched, isError } = useGeostoryLayers({ geostory_id });
   const { data: monitors } = useMonitors();
   const monitor = findMonitorByGeoStoryId(monitors, geostory_id);
+  const [layers] = useSyncLayersSettings();
+  const [compareLayers, setCompareLayers] = useSyncCompareLayersSettings();
 
+  const opacity = layers?.[0]?.opacity;
   const { title: monitorTitle, id: monitorId, color } = monitor || {};
+
+  // Only show layers with position right
+  const geostoryLayers = data?.filter(({ position }) => position === 'right');
+
+  useEffect(() => {
+    const comparisonLayer = data?.find(({ position }) => position === 'left');
+    if (comparisonLayer && !compareLayers) {
+      void setCompareLayers([{ id: comparisonLayer.layer_id, opacity }]);
+    }
+  }, [compareLayers, data, opacity, setCompareLayers]);
 
   return (
     <div className="space-y-6">
@@ -49,9 +65,14 @@ const GeostoryPage: NextPage<{ params: { geostory_id: string } }> = ({
 
         {isFetched && !isError && (
           <ul className="space-y-6" data-testid="datasets-list">
-            {data.map((dataset, index) => (
+            {geostoryLayers.map((dataset, index) => (
               <li key={dataset.layer_id}>
-                <DatasetCard {...dataset} id={dataset.layer_id} defaultActive={index === 0} />
+                <DatasetCard
+                  {...dataset}
+                  type="geostory"
+                  id={dataset.layer_id}
+                  defaultActive={index === 0}
+                />
               </li>
             ))}
           </ul>